Backend pulido

+++ Vantiq/Models/Categoria.cs
@@ -3,15 +3,15 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Vantiq.Models
 {
-    /// <summary>Categoria del reloj: Automatico, GMT, Cronografo.</summary>
+    /// <summary>Categoría del reloj: Automático, GMT, Cronógrafo.</summary>
     [Table("CATEGORIA")]
     public class Categoria
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public ushort IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
 
         [Required, StringLength(50)]
-        public string NombreCategoria { get; set; } = null!;    // UNICO
+        public string NombreCategoria { get; set; } = null!;    // ÚNICO
 
         [StringLength(150)]
         public string? Descripcion { get; set; }
